Introduce switching to dword I/O mode
After S_RESET the controller will be in the word i/o mode [1 p.101].
We need to perform 32bit write access to the i/o location at 0x10 to
switch the controller to dword i/o mode [1 p.104]. Zero is safe
value for this operation [1 p.104].
DWIO bit of BCR18 indicates that the controller is working in dword
mode. It should be set.

[1] Am79C970A